+++ Engine/Fire.cpp
@@ -2,18 +2,18 @@
 #include "Player.h"
 #include "Graphics.h"
 
-void Fire::DrawFire(int x,int y,Graphics &gfx)
+void Fire::DrawFire(float x,float y,Graphics &gfx)
 {
 	if (firebor == false)
 	{
-		gfx.PutPixel(x, -5 + y, 255, 255, 255);
-		gfx.PutPixel(x, -4 + y, 255, 255, 255);
-		gfx.PutPixel(x, -3 + y, 255, 255, 255);
+		gfx.PutPixel((int)x,-5 + (int)y, 255, 255, 255);
+		gfx.PutPixel((int)x,-4 + (int)y, 255, 255, 255);
+		gfx.PutPixel((int)x,-3 + (int)y, 255, 255, 255);
 	}
 }
 bool Fire::Border_Collide(Graphics &gfx)
 {
-	if (firey - 5< 0)
+	if (firey - (float)5< 0)
 	{
 		return  firebor = true;
 	}
@@ -21,12 +21,12 @@ bool Fire::Border_Collide(Graphics &gfx)
 		return firebor = false;
 }
 
-int Fire::GetFx()
+float Fire::GetFx()
 {
 	return firex;
 }
 
-int Fire::GetFy()
+float Fire::GetFy()
 {
 	return firey;
 }
@@ -36,12 +36,12 @@ bool Fire::GetBor()
 	return firebor;
 }
 
-void Fire::SetFx(int fx)
+void Fire::SetFx(float fx)
 {
 	firex = fx;
 }
 
-void Fire::SetFy(int fy)
+void Fire::SetFy(float fy)
 {
 	firey = fy;
 }
