+++ Engine/Object.h
@@ -5,20 +5,20 @@
 class Object
 {
 public:
-	void Init(int in_x, int in_y, int in_vx);
+	void Init(float in_x, float in_y, float in_vx);
 	void Object_Collide(Fire &fire);
 	void Update(Graphics &gfx);
-	void Border_Collide(int &x, int &y, int &vx,Graphics &gfx);
+	void Border_Collide(float &x, float &y, float &vx,Graphics &gfx);
 	void DrawBox(int,int, int, Graphics &gfx);
-	int GetOx();
-	int GetOy();
+	float GetOx();
+	float GetOy();
 	bool GetDes();
 	void SetDes(bool setdes);
 private:
-	int vx = 1;
-	int vy = 0;
-	int x =0;
-	int y= 0;
+	float vx = 1.0f;
+	float vy = 0.0f;
+	float x =0.0f;
+	float y= 0.0f;
 	bool collide = false;
 	bool Destroy = false;
